# force inlining of sequence iterator