@Ward9250 2016-08-26
Inlining methods frequently called in loops is generally a good idea in julia afaik :+1: