@bicycle1885 2016-08-26
But Julia 0.4 did it automatically IIRC, so I think this is a kind of regression in Julia.