@Ward9250 2016-08-26
That is curious, I usually did @inline on iterator methods I write anyway to be sure. I wonder why it doesn't do it automatically anymore.